// src/Types.ts
export interface Player {
  id: number;
  name: string;
  country: string;
  role: string;
  battingStyle: string;
  bowlingStyle: string;
  age: number;
  team: string;
  jerseyNumber: number;
  matches: number;
  runs: number;
  wickets: number;
  strikeRate: number;
  image: string;
  description: string;
  price: number;
}